requirements.txt
fastapi
uvicorn[standard]
pillow
pix2text
opencv-python-headless
python-multipart